Minor fixes. IO segfaults on quit when a directory (or file ?) is still referenced somewhere. Collecting files and dirs while running is fine, though.
git-svn-id: https://tinyaml.googlecode.com/svn/trunk@155 c1dc5971-5c3e-0410-84a0-c3311864b1d4

+++ src/tinyaml.c
@@ -255,7 +255,7 @@ long do_args(vm_t vm, long argc,char*argv[]) {
 
 
 
-long main(long argc, char** argv) {
+int main(int argc, char** argv) {
 	vm_t vm;
 	vm = vm_new();
 	/*vm_set_engine(vm, stub_engine);*/
